pyproject.toml
[project]
name = "curso-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.13",
    "chromadb>=1.5.5",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "openai>=2.30.0",
    "pymysql>=1.1.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.43.0",
]
